drivers: npcx: replace type of register base address with uintptr_t
Replace npcx register base address type, uint32_t, with uintptr_t.
It is easier to know what type of base address and for linear
addresses treated as integral values.

This CL also modified IS_BIT_SET() macro function to fit MISRA code
guidelines.

+++ drivers/clock_control/clock_control_npcx.c
@@ -16,9 +16,9 @@ LOG_MODULE_REGISTER(clock_control_npcx, LOG_LEVEL_ERR);
 /* Driver config */
 struct npcx_pcc_config {
 	/* cdcg device base address */
-	uint32_t base_cdcg;
+	uintptr_t base_cdcg;
 	/* pmc device base address */
-	uint32_t base_pmc;
+	uintptr_t base_pmc;
 };
 
 /* Driver convenience defines */
